简化类
