--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/eventloop/integral/StructuralLinearMode.java
@@ -1,13 +1,0 @@
-package org.firstinspires.ftc.teamcode.eventloop.integral;
-
-/**
- * 提供结构化方法
- */
-public abstract class StructuralLinearMode extends IntegralLinearMode {
-	@Override
-	public Thread getLinearThread() {
-		return new Thread(this::linear);
-	}
-
-	public abstract void linear();
-}
