+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/events/IntegralThreadExceptionHandler.java
@@ -10,7 +10,6 @@ import com.qualcomm.robotcore.eventloop.opmode.OpModeManagerImpl;
 
 import org.betastudio.ftc.Interfaces;
 import org.firstinspires.ftc.teamcode.eventloop.TerminateReason;
-import org.firstinspires.ftc.teamcode.eventloop.integral.IntegralAutonomous;
 import org.firstinspires.ftc.teamcode.eventloop.integral.IntegralOpMode;
 
 import java.lang.Thread.UncaughtExceptionHandler;
@@ -42,7 +41,7 @@ public class IntegralThreadExceptionHandler implements UncaughtExceptionHandler 
 			// 如果异常不是OpModeManagerImpl.ForceStopException类型，记录错误日志
 			Log.e("Error", "OpMode Terminated By Exception", e);
 			// 根据当前操作模式的类型，发送终止信号或直接终止操作模式
-			if (currentOpmode instanceof IntegralAutonomous) {
+			if (currentOpmode instanceof IntegralOpMode) {
 				((IntegralOpMode) currentOpmode).sendTerminateSignal(TerminateReason.UNCAUGHT_EXCEPTION, (Exception) e);
 			} else {
 				currentOpmode.terminateOpModeNow();
